requirements.txt
langchain
langchain-google-genai
langgraph
streamlit
python-dotenv
google-auth
google-api-python-client
gTTS
google-generativeai
langchain-huggingface
huggingface_hub
langchain-openai
